Renaming spotterThresholds<List> in Server.java to similarityThresholds<List>
This might lead to confusion. The spotter_thresholds.txt in the statistical backend are not the same as the similarityThresholds in the Lucene backend. Eventually, we might add similarityThresholds to the statistical backend, too. They are just a way of estimating confidence. The spotter_thresholds.txt in the statistical backend are weights for scores that go into deciding if a SF is annotated (spotter_weights.txt might actually be a better name for them)